Added icon generation and bug fix
If icon.svg exists in the root of the project folder and is newer than the
  platform icons, new icons will be generated automatically at buld time using
  inkscape.
Added if statements to hooks to prevent the errant creation of a www/www
  directory.

+++ 050_fivel_after_build.sh
@@ -1,4 +1,6 @@
 #!/bin/bash
 
-rm -rf www
-mv www~ www
+if [ -e www~ ]; then 
+  rm -rf www
+  mv www~ www
+fi;
